chore: move project details types to types folder
