+++ src/app/(main)/_components/project-details/action-buttons.tsx
@@ -2,8 +2,8 @@
 import { Button } from "@/components/ui/button";
 import { Plus } from "lucide-react";
 import Link from "next/link";
+import type { Projects } from "../../../../lib/types/project/project-details.types";
 import ExportPdfButton from "./export-pdf-button";
-import type { Projects } from "./interface";
 import ProjectInfoButton from "./project-info-button";
 import ShareButton from "./share-button";
