+++ src/app/(main)/_components/project-details/contract-details.tsx
@@ -1,4 +1,4 @@
-import type { Projects } from "./interface";
+import type { Projects } from "../../../../lib/types/project/project-details.types";
 export default function ProjectContractDetails({
     project,
 }: {
