+++ src/app/(main)/_components/project-details/project-info-button.tsx
@@ -17,7 +17,7 @@ import {
     TriangleAlert,
 } from "lucide-react";
 import { useState } from "react";
-import type { Projects } from "./interface";
+import type { Projects } from "../../../../lib/types/project/project-details.types";
 
 export default function ProjectInfoButton({ project }: { project: Projects }) {
     const [isDetailsOpen, setIsDetailsOpen] = useState(false);
