+++ src/lib/types/project/project-details.types.tsx
